[ui] Fix events missing on Asset > Events when mat + observations both present (#25713)
## Summary & Motivation

The previous fix I wrote for
https://linear.app/dagster-labs/issue/FE-387/assets-with-both-materializations-and-observations-can-have-plot
in #25416 put the "clipping to a shared time range" logic in
`useGroupedEvents`, which had an unintended side-effect I noticed this
morning.

This code wasn't verifying that one of the datasets contained `limit`
items before clipping them both to the shared timerange. This meant that
if you had an asset with a history of materializations, and then emitted
an observation for it, the older materializations would disappear from
the asset plots. More concerningly they also disappeared from the Asset
> Evetns page, which was using `useGroupedEvents` because it can perform
grouping on super old non-SDAs that don't have a defined parititon
space.

I moved the logic out into a helper and wrote tests for it, and then
made both `useRecentAssetEvents` and `usePaginatedEvents` use the same
logic. (`usePaginatedEvents` isn't used for plots, but needs the same
logic because it uses a single pagination cursor to move through asset
observations and materializations at the same time, so the "next" cursor
must be the greater of the two minimum values).

Long term, it would be ideal to have a single `events` resolver that
returned both materializations and observations, because fetching them
separately is proving to be awkward.

## How I Tested These Changes

New test cases + manual verification with 1) assets that have a long
history of mixed obs + materialization events, and 2) an asset that has
a history materializations and just one new observation.

## Changelog

[ui] The Asset > Events page no longer truncates event history in cases
where both materialization and observation events are present.

Co-authored-by: bengotow <bgotow@elementl.com>