[ui] Fix poor performance when selecting assets with many partitions (#25718)
## Summary & Motivation

I was testing another bug and noticed that selecting assets in the asset
graph with 200k+ partitions was freezing the UI for 30+ seconds. I dug
into this and found this component was still iterating over the
partition space to find the number of materialized partitions. This is
now available via `liveData.partitionStats`, and the component is only
used in one place with no `selections` passed in.

I think this code used to be necessary because the component was also
used in other contexts where there were either 1) multiple assets being
displayed as a single health bar or 2) a subset of the partition space
in `selections`, but with just this one callsite left there's no reason
to not just show the counts from the asset health data.

## How I Tested These Changes

Tested manually with large assets that load slowly. Clicking this asset
is now fast:
<img width="1358" alt="image"
src="https://github.com/user-attachments/assets/84b736e2-be54-436e-a65b-39b1d064fdce">


## Changelog

[ui] Selecting assets with 100k+ partitions no longer causes the asset
graph UI to stutter

Co-authored-by: bengotow <bgotow@elementl.com>