[2/n][dagster-fivetran] Update DagsterFivetranTranslator and related classes for rework (#25751)
## Summary & Motivation

~~Builds out a very barebones translator class for the new version of
the Fivetran integration.~~

~~The implementation for this translator will be inspired by the
`DagsterFivetranTranslator` introduced in
https://github.com/dagster-io/dagster/pull/25557, but a new
implementation is required to leverage the workspace context and
state-backed definitions, which is incompatible with the current
translator and way of building assets.~~

Edit after Ben's comment
[here](https://github.com/dagster-io/dagster/pull/25751#discussion_r1830086548):

Move things around under translator.py. This PR leverages the
`DagsterFivetranTranslator` introduced in introduced in
https://github.com/dagster-io/dagster/pull/25557.
`FivetranWorkspaceData` will implement the method
`to_fivetran_connector_table_props_data` in a subsequent PR, that will
map raw connector and destination data fetched using the Fivetran API
into `FivetranConnectorTableProps` objects, that are compatible with the
translator. This process will match what we currently do.

## How I Tested These Changes

Tests will be added in subsequent PRs.