[run status sensor] allow matching None origins for testing (#25795)
Ffor testing `run_status_sensor` we currently only support using
`build_run_status_sensor_context` and directly invoking the sensor to
exercise just the decorated function. If a user wants to ensure that
they have all of the arguments set correctly on their status sensors and
that they react to an `instance` as expected when events happen they
currently can not.

This PR makes a small tweak to the code origin match logic to match when
both are `None` to allow in process execution matches. This makes it so
it as at least possible to use `evaluate_tick` on a `@run_status_sensor`
with in process execution to test how it would react.

Theres more that could be done to make this easier, but this at least
makes it possible.

## How I Tested These Changes

added test