[dagster-pipes] arbitrary metadata for PipesClientCompletedInvocation (#24342)
## Summary & Motivation

Currently `Pipes` clients only return information and metadata from the external process side and there is no way to add information only available on the orchestration side.

This PR allows passing metadata to `PipesClientCompletedInvocation` invocation. This metadata will be attached to all Pipes materializations (explicit or implicit).  

## How I Tested These Changes

Added tests

## Changelog

Added the ability to pass arbitrary metadata to `PipesClientCompletedInvocation`. This metadata will be attached to all materializations and asset checks stored in this object.