Refactor AssetOut to hold AssetSpec (#25789)
## Summary

`AssetOut` is a bit of an unfortunate API in its current state - it's a
huge `NamedTuple` which holds a large subset of the fields of
`AssetSpec` along with some extra fields. It can be converted to an
`AssetSpec` - in which case we manually construct a spec passing all
those parameters.

This PR attempts to remedy this a bit by backing `AssetOut` internally
with an `AssetSpec` - meaning the `to_spec` implementation can simply
forward this spec, and also letting us construct an `AssetOut` using a
spec. This lets users e.g. pass `kinds` which don't have a matching
param to `AssetOut`:

```python
@graph_multi_asset(
    outs={
        "out1": AssetOut.from_spec(AssetSpec(key="my_key", kinds={"python", "s3"})),
        "out2": AssetOut.from_spec(
            AssetSpec(key="my_other_key", kinds={"python", "snowflake"})
        ),
    }
)
def my_graph():
    return layered_op(inner_op()), layered_op(inner_op())
```

## Open Questions

- Should we just allow `AssetSpec` to be passed in to `outs`? We lose
the extra params that `AssetOut` provides in this case, but for users
who aren't using them it might be a cleaner option.
- Would we prefer `AssetSpec(...).as_out(...)`


## Test Plan

New unit test, existing tests.

## Changelog

> Introduced `AssetOut.from_spec` to construct an asset out from a spec