Make 'multiple Definitions objects' error message more actionable (#25453)
## Summary

Ran into this with a user - the "Cannot have more than one Definitions
object defined at module scope." error can be a bit of a rabbithole to
track down. The case we hit in particular was
when one of the `Definitions` objects was being imported unknowingly.

This PR updates the error message to include the names of all the
located `Definitions` objects, so it's easier to track down where
they're coming from and rectify the error.

## How I Tested These Changes

Update unit tests.

## Changelog

> Errors raised from defining more than one `Definitions` object at
module scope now include the object names so that the source is easier
to determine.