Add buffer time to perf test (#25915)
## Summary & Motivation

As title. This is a consequence of getting rid of the special-purpose
PartitionKeysTimeWindowPartitionsSubset. Long story short, these perf
tests are not very representative of real-world perf scenarios, and
mostly exist as a backstop in case anything massively changes. This
change is also completely irrelevant to the Declarative Automation perf
characteristics, which are more important and have their own separate
perf tests.

TL;DR: We'll probably remove this test in the near future, for now we're
just making it stop failing

## How I Tested These Changes

## Changelog

NOCHANGELOG