[dagster-airlift][federation-apis] federation APIs accumulator (#25857)
This PR is now being used to accumulate federation API commits before landing.
ORIGINAL DESCRIPTION:
## Summary & Motivation
Right now, all airlift internals which operate over mapped airflow assets take in a fully-qualified Definitions object. This PR changes them to instead operate over an iterable of assets, and leaves the definitions munging to the callsites.

After making this change, it becomes easy to add asset-oriented airlift methods - and I demonstrate that with a standalone method to enrich airflow assets. We'll eventually be able to re-implement the full defs loader on top of this.

## How I Tested These Changes
New test for the new method, existing test suite.

## Changelog
NOCHANGELOG