Get rid of risk of order dependence (#25943)
## Summary & Motivation

Noticed this while working in this area -- the LoadingContext system expects the output of these methods to be in the same order as the input, which is not currently guaranteed. The AssetStatusCacheValues don't have a reference to what asset key they came from, meaning we can't reconstruct this mapping from an unordered list.

Simplest solution is to just not use dictionaries at all.

Surprised this hasn't caused issues, to be honest!

Internal: https://github.com/dagster-io/internal/pull/12702

## How I Tested These Changes

## Changelog

NOCHANGELOG