avoid fetching runs in runs feed if backfill filter is set and we are hiding runs within backfills (#25904)
## Summary & Motivation
If a user filters for a specific backfill in the default runs feed view,
we end up making an inefficient query like this
```sql
SELECT * FROM runs
WHERE runs.backfill_id IS NULL
AND runs.backfill_id = 'abcdefgh'
```

This is because the default view of the Runs page excludes subruns,
which is the `runs.backfill_id IS NULL`. This is an inefficient query
and we have enough knowledge prior to running the query to know it will
return no results. So we can just skip fetching runs if we see we are in
this case

## How I Tested These Changes
new unit test