Fix issue where ToposortedPriorityQueue could return children after their parents
Summary:
Doing this max could cause us to break the invariant that parents are always considered before parents.

If you have a graph like this:

```

@asset
def a():
    pass

@asset
def b(a):
    pass

@asset
def c(b):
    pass

@multi_asset(specs=[AssetSpec("d", deps=["a"]), AssetSpec("e", deps=["c"])])
def my_complex_assets():
    pass

@asset(deps=["d"])
def f():
    pass
```

and pass in both a and f, it would evaluate f before d, because d gets incorrectly bumped up to the same toposort level as e.

Now, instead, D+E will always be visited before F. This arguably still breaks the invariant in a different way, in that E can now be visited (as part of the D+E multi-asset) before C, but I think that is happening in a more principled way now because when that happens D is still the 'main' asset and E is just along for the ride since it's part of D's multi-asset

It looks like this was added way back in https://github.com/dagster-io/dagster/pull/11190/files when this logic was being used for AMP, which appears to no longer be the case. Now it is only used in asset backfills to determine the full set of eligible asset partitions (where it is actually no longer clear to me that it is a hard requirement that parents always be evaluated before their children, as long as everything gets evaluated eventually)

> Insert changelog entry or delete this section.