[ui] Lint against window.open (#25912)
## Summary & Motivation

Stacking on https://github.com/dagster-io/dagster/pull/25909, lint
against using `window.open`. This is because our current usage generally
ignores any `basePath` value that might be in place in the app, e.g. a
deployment path prefix in Dagster+.

The result is that opening the new tab may lead to a redirect that opens
the path in a default fallback deployment, instead of the viewer's
current deployment.

To avoid this, `useOpenInNewTab` prepends the basePath to the path
value. I have added a lint rule to error on `window.open` usage and push
users toward this hook instead.

## How I Tested These Changes

TS, lint, jest.