ANTLR only allow traversal on certain expressions (#26068)
## Summary & Motivation
We discovered some ambiguity caused by traversal expressions that may cause parser to return an unexpected asset selection. For example, currently the parser reads `*key:a and key:b*` as `*(key:a and key:b)*`. We believe this is unintuitive, so we've decided to force the traversal tokens to only apply to the expression they are immediately next to.