[ui] Open evaluation dialog from sensor page (#26079)
## Summary & Motivation

When viewing automaterialization evaluations from a sensor page, open the evaluation details in a dialog instead of the asset's Automation tab. This means opening a dialog on top of another dialog, instead of navigating away.

This change is in preparation for updating the Automation tab to show the list of evaluations as the main content, which will involve opening evaluation details in a dialog as well.

https://github.com/user-attachments/assets/4adca993-6463-4aec-85d5-b4f0bf37dfef

## How I Tested These Changes

View an enabled automation sensor. View materializations in dialog, then view individual evaluations. Verify that the dialog opens and renders properly, and that partitioned assets behave correctly when choosing partitions within the dialog.

View Automation tab on individual assets, verify that these remain unaffected and behave the same as before.

## Changelog

[ui] Open automation sensor evaluation details in a dialog instead of navigating away.