Do not collapse `AutomationCondition`s with labels (#26097)
## Summary & Motivation

Imagine you have:

```python
a = (AutomationCondition.foo() & AutomationCondition.bar()).with_label("something")
b = a & AutomationCondition.baz()
```

The implementation of `__and__` automatically transforms `And(And(c1,
c2), c3)` into `And(c1, c2, c3)`. This process is generally desirable,
as this reduces the number of levels in the hierarchy, making it easier
to parse and understand expressions.

However, if an expression has a label, then this indicates that it has
its own semantic meaning, which gets erased by this process.

We should prevent this automatic collapsing in cases where the existing
condition's label is not `None`.

The same applies to `__or__`!

## How I Tested These Changes

Added unit tests in
python_modules/dagster/dagster_tests/definitions_tests/declarative_automation_tests/automation_condition_tests/fundamentals/test_automation_condition.py.

## Changelog

Fixed a bug where using the `&` or `|` operators on
`AutomationCondition`s with labels would cause that label to be erased.