[ui] Stop using the job page as the home page when single job is defined (#26120)
## Summary & Motivation


https://linear.app/dagster-labs/issue/FE-683/change-fallthrough-redirect-not-to-land-on-single-job

This case is somewhat rare -- if a user has exactly one job, we had
special cased it and opening the Dagster UI would land you directly on
the job page. We think that this case is odd and it'd be better to
direct to the Overview.

## How I Tested These Changes

I added test coverage for all variants of the landing page behavior. 🫡 

## Changelog

[ui] Opening Dagster's UI with a single job defined takes you to the
Overview page rather than the Job page

Co-authored-by: bengotow <bgotow@elementl.com>