[components] Initial implementation of component project scaffolding (#26104)
## Summary & Motivation

This is an initial implementation of scaffolding for components. It adds
a new CLI `dg` rather than a subcommand to `dagster`. There is other
work in flight but this seemed like a good place to cut the bottom of
the stack. Implements:

- `dg generate deployment`
- `dg generate code_location`
- `dg generate component-type`
- `dg generate component-instance`

What is NOT in this PR:

- global component registry based on entrypoints-- instead this only
resolves component types local to the code location
- anything resembling final `Component` API. This uses a very bare-bones
prototype implementing `build_defs` and `generate_files` that will be
replaced.
- Calling out to `uv` in scaffolding commands.
- Thoroughly polished output for scaffold commands.
- Final names for commands (`component-type`, `component-instance` are
questionable)

All of ^^ will come upstack or from @OwenKephart efforts.

Note that it was necessary to expose `Component` from top-level
`dagster` so that realistic generated files could be used (`from dagster
import Component`). To avoid exposing this to users I added it as a
dynamic import.

Note also there is some weirdness with `sys.path`, which by default has
`""` as first entry (allowing imports from cwd) but which for some
reason is missing this entry in the test environment. I've hacked it in
with some code that will probably go away once the underyling issue can
be identified.

## How I Tested These Changes

New unit tests.