[components] Basic PythonScript component implementation (#26143)
## Summary & Motivation

This stands up the python API for instantiating a component, and provides a sample implementation. Upstack, utilities are provided for instantiating a component from a config blob.

Also provides a ComponentCollection class, which is a generic way to encapsulate multiple components of the same type. This becomes more useful in the config-based instantiation paths.

## How I Tested These Changes

## Changelog

NOCHANGELOG