[components] Add ability to load a component using config (#26148)
## Summary & Motivation

This adds a separate construction path for Components, which takes a dictionary of config that conforms to the class's config schema.

Because there's no generalized way to auto-generate a config schema or a conversion function, I think we'll end up relying in most cases on users subclassing an iniital simple implementation that just allows for some configured specs and maybe some unstrctured params, hence the AssetsComponent.

## How I Tested These Changes

## Changelog

> Insert changelog entry or delete this section.