[components] Remove nested component concept, pass path directly through init_context (#26180)
## Summary & Motivation

This removes the ComponentCollection concept, and replaces it with a FileCollectionComponent, which is just a subclass that makes it easier to do per-file operations.

My feeling was that the ComponentCollection thing was over-specified, and added a weird layer of indirection to the system. As can be seen in the diff, this change made implementation a fair amount simpler for a variety of things, while not making the PythonScript component implementation much different.

The key thing here is that it feels very unlikely that someone would want to use a given component type both within and without of a Collection wrapper (i.e. if it's designed to go inside a Collection, it will always be in a collection). If that's the case then we should just have them define the collection component directly.

## How I Tested These Changes