[components] Change registered component name to just Component.registered_name (#26175)
## Summary & Motivation

Prior to this PR, component types would get registered as
`module.path[component-type-name]`. This PR changes it to just
`component-type-name`. I don't think this is a final decision, but this
is a useful simplification for now since it allows us to punt on
decisions about the module name for locally defined components.

## How I Tested These Changes

Existing test suite.