[pipes] add PipesStdioLogWriter (#25764)
## Summary & Motivation

This PR adds:

- `PipesStdioLogWriter` - an partial implementation for `PipesLogWriter` abstract class which captures stdout/err of the current process. It doesn't handle writing logs, only capturing. Child classes should implement writing.
- `PipesStdioFileLogWriter` - an implementation for `PipesStdioLogWriter` which writes logs to the local filesystem (to a given directory)

---

The solution relies on `tee` command so only works on Linux and MacOS. This is required for real-time logs capturing because `tee` creates a normal file where stdout/stderr is duplicated, and it's possible to read from this file immediately (in oppose to other methods such as just running `os.dup2`).  

## How I Tested These Changes

Added tests

## How I Tested These Changes

Added a test. Tests for the Dagster side are upstream.

## Changelog

No changelog