[components] Explicit schema for PipesSubprocessScriptCollection (#26216)
## Summary & Motivation

This changes `PipesSubprocessScriptCollection` to rely less on automatic registration of scripts and instead relies on more explicit configuration. The behavior was really unexpected and there is no way to opt out of it (e.g. if there is a shared python file for common code).

I think we re-add some autodiscovery capabilities but built on this foundation.

## How I Tested These Changes

BK