Basic SlingReplicationComponent (#26260)
## Summary & Motivation

This implements a basic SlingReplicationComponent, which does a few things:

1. On generation, creates a mostly-empty `replication.yaml` file
2. Uses the existing sling_assets decorator to build an AssetsDefinition based on this replication.yaml file
3. Configures connection info based off of information stored in the `defs.yaml` file

This final bit is the first toe in the water of resource configuration handling. In this case, I just embed the resource directly into the config schema, which works out quite nicely in practice.

## How I Tested These Changes

It's a bit tricky to set up tests for this, as we need to create a temporary directory so that there's an absolute path that we can reference in the stream config. Set up a fixture for this purpose

## Changelog

NOCHANGELOG