[DA] Update docs (#26262)
## Summary & Motivation

This splits up the existing customizing-automation-conditions page into
two. The first is a slimmed-down version of the original page that stays
much more focused on specific use cases. I also took the liberty of
backing out of the original sin of directly embedding python into the
mdx (now it's broken out into separate python files, at least for the
new stuff).

The second page is a more advanced conceptual guide for the ins and outs
of AutomationConditions. This is not necessary information for the
majority of users, but I think it is valuable to have in some place to
give interested users a more detailed conceptual understanding of how
the system fits together, especially if they have highly-specific use
cases that require deep manipulation of the APIs.

Also, some of the existing code examples in the
cusotmizing-automation-conditions page are a bit complex, and will serve
as good motivating examples for some lightweight chagnes to the core
APIs to make those examples simpler


## How I Tested These Changes

## Changelog

NOCHANGELOG