[backfill daemon run retries 2/n] backfill daemon incorporates retries runs when launching new runs (#25853)
## Summary & Motivation
The backfill daemon doesn't account for run retries. See
https://github.com/dagster-io/internal/discussions/12460 for more
context. We've decided that we want the daemon to account for automatic
and manual retries of runs that occur while the backfill is still in
progress. This requires two changes: ensuring the backfill isn't marked
completed if there is an in progress run or a failed run that will be
automatically retried; and updating the daemon to take the results of
retried runs into account when deciding what partitions to materialize
in the next iteration.

This PR addresses the second point, updating the backfill daemon to take
the results of retried runs into account when deciding what partitions
to materialize in the next iteration.

Currently the backfill gets a list of the successfully materialized
assets for the backfill by looking at the materialization events for the
asset. It determines which assets failed by looking at the failed runs
launched by the backfill and pulling the asset partition information
from those runs. Any assets downstream of those failed assets will not
be launched by the backfill

Now that we want the backfill daemon to account for run retries we need
to slightly modify this logic. Since a run can be retried it is possible
that an asset can have a successful materialization AND be a failed
asset in a failed run. This means that when we determine which assets
are failed, we need to cross check with the assets that have been
successfully materialized and remove any that are in the materialized
list

## How I Tested These Changes
new unit tests 

manual test 

```python
@dg.asset(
    partitions_def=daily_partitions_def,
)
def has_retry_policy(context):
    if context.run.parent_run_id is None:
        raise Exception("I failed")
    else:
        return 1

@dg.asset(partitions_def=weekly_partitions_def)
def another_weekly(context, has_retry_policy):
    return 1
```

run a backfill of at least 1 week. Can watch the backfill progress and
see that once all of the daily assets retry and succeed, the weekly
asset will get kicked off. This contrasts with the behavior in the PR
down the stack where the daily assets would retry and succeed, but the
weekly asset would never get kicked off

## Changelog

Manual and automatic retries of runs launched by backfills that occur
while the backfill is still in progress are now incorporated into the
backfill's status