[backfill daemon run retries 3/n] retries of runs in completed backfills should not be considered part of the backfill (#25900)
## Summary & Motivation
If a run is retried after a backfill is complete, that run is given the
backfill tag, but has no affect on the backfill itself. This can cause
confusion. Imagine the scenario where a single asset-partition failed in
a backfill. The backfill is complete and a user retries the failed asset
and the retry succeeds. That retried run will show up in the list of
runs for the backfill, but the status in the overview tab for partition
will still be failed since the status is locked when the backfill
completes.

We should be more strict about when run retries are considered part of
the backfill. We decided in
https://github.com/dagster-io/internal/discussions/12460 that retries
that are launched while the backfill is in progress will be part of the
backfill, but that retries that are launched after the backfill is
complete should not be considered part of the backfill.

To make this change we need to remove the backfill tag from retried runs
if the backfill is not in progress.

## How I Tested These Changes
new unit tests 

manually launched a retry of a run that was launched by a backfill after
the backfill was complete. no backfill tags were added
<img width="1037" alt="Screenshot 2024-12-02 at 1 55 09 PM"
src="https://github.com/user-attachments/assets/5bb8ae12-4c61-4fd4-8255-1d245ae43318">


## Changelog

Manual retries of runs launched by backfills are no longer considered
part of the backfill if the backfill is complete when the retry is
launched.