[dagster_aws][pipes] Don't cause an error when get_log_events raises certain exceptions
When triggering an ECS task it usually takes 30s plus for the container
to start up, and so the log stream typically isn't available for that
period.

Currently we attempt to get_log_events() before the log stream is
available which causes errors to show on Dagster.

This change catches the errors so they don't show. They will still
appear in stderr but not in the Events output. This also fixes the
result of the run showing on the materialized asset at the end of the
run, where as previously the thrown error would stop it receiving the
required messages back from the ECS task.

Simlarly, sometimes ThrottlingException's are thrown if you have a few
tasks on the go querying logs. This will handle this as well so it
doesnt' pollute the event view and break the result materialization.
Again, these errors are still shown in stderr.